moved folders to the src directory
I think this layout is a bit cleaner, grouping app, components etc together to keep the root a little cleaner
