+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 
 import { siteConfig } from 'config/site';
-import { buttonVariants } from 'components/ui/button';
+import { buttonVariants } from '@/src/components/ui/button';
 
 export default function IndexPage() {
   return (
