+++ src/components/nav/site-header.tsx
@@ -1,10 +1,10 @@
 import Link from 'next/link';
 
 import { siteConfig } from 'config/site';
-import { buttonVariants } from 'components/ui/button';
-import { Icons } from 'components/icons';
-import { MainNav } from '@/components/nav/main-nav';
-import { ThemeToggle } from 'components/theme-toggle';
+import { buttonVariants } from '@/src/components/ui/button';
+import { Icons } from '@/src/components/icons';
+import { MainNav } from '@/src/components/nav/main-nav';
+import { ThemeToggle } from '@/src/components/theme-toggle';
 
 export function SiteHeader() {
   return (
