+++ tailwind.config.js
@@ -3,7 +3,7 @@ const { fontFamily } = require('tailwindcss/defaultTheme');
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   darkMode: ['class'],
-  content: ['app/**/*.{ts,tsx}', 'components/**/*.{ts,tsx}'],
+  content: ['src/app/**/*.{ts,tsx}', 'src/components/**/*.{ts,tsx}'],
   theme: {
     container: {
       center: true,
